Improve user comparisons
Now user comaprisons calculates language similarities (instead of differences),
and we think this is a more accurate way to compare against.

The activity difference comparison tries to compare based on mean difference
percentages, we expect that by using percentages we will be able to cuantify how
differente these values are.

In cases where mean is not useful (mean is 0), then we use the standard
deviation percentage, and compare how different their variability is.